refactoring(styles): replacing style files
Made-with: Cursor
