+++ src/components/ConstructorComponent/ConstructorComponent.scss
@@ -1,4 +1,4 @@
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 $sidebarWidth: 384px; // 768 / 2 MediumDevices and higher
 $paddingtopForCloseButton: 50px;
