+++ src/components/CountdownEditor/CountdownEditor.scss
@@ -1,4 +1,4 @@
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 .CountdownEditor {
     position: relative;
