+++ src/components/GridSlot/GridSlot.scss
@@ -1,4 +1,4 @@
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 .EmptySlot {
   width: 100%;
