+++ src/components/Header/header.scss
@@ -1,5 +1,5 @@
-@use '../../variables.scss' as vars;
-@use '../../mixin.scss' as *;
+@use '../../styles/variables.scss' as vars;
+@use '../../styles/mixin.scss' as *;
 
 .header {
     display: flex;
