+++ src/components/HotkeyCell/HotkeyCell.scss
@@ -1,4 +1,4 @@
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 .HotkeyCellWrapper {
     height: 38px;
