+++ src/components/LangSelector/LangSelector.scss
@@ -1,4 +1,4 @@
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 $menuWidth: 170px;
 $menuHeadHeight: 40px;
