+++ src/components/LevelController/LevelController.scss
@@ -1,4 +1,4 @@
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 .LevelController {
     display: flex;
