+++ src/components/Notification/Notification.scss
@@ -1,4 +1,4 @@
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 .Notification {
     width: 630px;
