+++ src/components/PlayButton/PlayButton.scss
@@ -1,4 +1,4 @@
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 $buttonHeight: 80; // px
 $buttonWidth: 380; // px
