+++ src/components/PlayButton/SmallPlayButton.scss
@@ -1,5 +1,5 @@
 @use 'sass:color';
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 $smallButtonHeight: 40; // px
 $smallButtonWidth: 173; // px
