+++ src/components/Timer/Timer.scss
@@ -1,4 +1,4 @@
-@use '../../variables.scss' as vars;
+@use '../../styles/variables.scss' as vars;
 
 .Timer {
     width: 100%;
