+++ src/pages/about/AboutPage/AboutPage.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {useSelector} from 'react-redux';
 import {TStoreState} from '../../../store/store';
 import { translateText } from '../../../utils/utils';
-import '../../../article.scss';
+import '../../../styles/article.scss';
 import ImageCover from '../../../components/ImageCover/ImageCover';
 import SpriteIcon from '@shared/ui/SpriteIcon';
