+++ src/pages/contact/ContactPage/ContactPage.tsx
@@ -3,7 +3,7 @@ import Social from '../../ui/social';
 import {useSelector} from 'react-redux';
 import {TStoreState} from '../../../store/store';
 import { translateText } from '../../../utils/utils';
-import '../../../article.scss';
+import '../../../styles/article.scss';
 import ImageCover from '../../../components/ImageCover/ImageCover';
 import SpriteIcon from '@shared/ui/SpriteIcon';
