+++ src/pages/donation/ui/qr-component.scss
@@ -1,4 +1,4 @@
-@use '../../../variables.scss' as vars;
+@use '../../../styles/variables.scss' as vars;
 
 .QrComponent {
     position: absolute;
