+++ src/pages/playground/PlaygroundPage/SettingsStage/SettingsStage.scss
@@ -1,4 +1,4 @@
-@use '../../../../variables.scss' as vars;
+@use '../../../../styles/variables.scss' as vars;
 
 .SettingsStage {
     width: 100%;
