+++ src/pages/welcome/WelcomePage/welcome.scss
@@ -1,5 +1,5 @@
-@use '../../../variables.scss' as vars;
-@use '../../../mixin' as *;
+@use '../../../styles/variables.scss' as vars;
+@use '../../../styles/mixin' as *;
 
 .Welcome {
     height: 100vh;
